+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparameter/snippet/forkononnullabletypeprovider/class-has-complex-default-parameter-value.kttxt
