+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/koparameter/snippet/forkononnullabletypeprovider/class-has-one-parameter-with-import-alias.kttxt
