+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kotype/KoTypeDeclarationForKoNullableProviderTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments.arguments
 import org.junit.jupiter.params.provider.MethodSource
 
-class KoTypeDeclarationForKoNullableTypeProviderTest {
+class KoTypeDeclarationForKoNullableProviderTest {
     @ParameterizedTest
     @MethodSource("provideValues")
     fun `is-nullable`(
