+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kotypealias/snippet/forkononnullabletypeprovider/typealias-has-lambda-type.kttxt
