+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kovariable/KoVariableDeclarationForKoNullableTypeProviderTest.kt
@@ -6,7 +6,7 @@ import org.amshove.kluent.assertSoftly
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 
-class KoVariableDeclarationForKoPropertyTypeProviderTest {
+class KoVariableDeclarationForKoNullableTypeProviderTest {
     @Test
     fun `variable-without-type`() {
         // given
@@ -47,5 +47,5 @@ class KoVariableDeclarationForKoPropertyTypeProviderTest {
     }
 
     private fun getSnippetFile(fileName: String) =
-        TestSnippetProvider.getSnippetKoScope("core/declaration/kovariable/snippet/forkopropertytypeprovider/", fileName)
+        TestSnippetProvider.getSnippetKoScope("core/declaration/kovariable/snippet/forkonullabletypeprovider/", fileName)
 }
