+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kovariable/snippet/forkonullabletypeprovider/variable-with-type.kttxt
