+++ lib/src/integrationTest/kotlin/com/lemonappdev/konsist/core/declaration/kovariable/snippet/forkonullabletypeprovider/variable-without-type.kttxt
