+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/declaration/KoPropertyDeclaration.kt
@@ -13,9 +13,9 @@ import com.lemonappdev.konsist.api.provider.KoKDocProvider
 import com.lemonappdev.konsist.api.provider.KoLocationProvider
 import com.lemonappdev.konsist.api.provider.KoModuleProvider
 import com.lemonappdev.konsist.api.provider.KoNameProvider
+import com.lemonappdev.konsist.api.provider.KoNullableTypeProvider
 import com.lemonappdev.konsist.api.provider.KoPackageProvider
 import com.lemonappdev.konsist.api.provider.KoPathProvider
-import com.lemonappdev.konsist.api.provider.KoPropertyTypeProvider
 import com.lemonappdev.konsist.api.provider.KoReceiverTypeProvider
 import com.lemonappdev.konsist.api.provider.KoResideInPackageProvider
 import com.lemonappdev.konsist.api.provider.KoSetterProvider
@@ -46,7 +46,7 @@ interface KoPropertyDeclaration :
     KoConstructorDefinedProvider,
     KoContainingFileProvider,
     KoDelegateProvider,
-    KoPropertyTypeProvider,
+    KoNullableTypeProvider,
     KoFullyQualifiedNameProvider,
     KoInitializerProvider,
     KoKDocProvider,
