+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/declaration/KoTypeAliasDeclaration.kt
@@ -9,12 +9,12 @@ import com.lemonappdev.konsist.api.provider.KoKDocProvider
 import com.lemonappdev.konsist.api.provider.KoLocationProvider
 import com.lemonappdev.konsist.api.provider.KoModuleProvider
 import com.lemonappdev.konsist.api.provider.KoNameProvider
+import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
 import com.lemonappdev.konsist.api.provider.KoPackageProvider
 import com.lemonappdev.konsist.api.provider.KoPathProvider
 import com.lemonappdev.konsist.api.provider.KoResideInPackageProvider
 import com.lemonappdev.konsist.api.provider.KoSourceSetProvider
 import com.lemonappdev.konsist.api.provider.KoTextProvider
-import com.lemonappdev.konsist.api.provider.KoTypeProvider
 import com.lemonappdev.konsist.api.provider.modifier.KoActualModifierProvider
 import com.lemonappdev.konsist.api.provider.modifier.KoModifierProvider
 import com.lemonappdev.konsist.api.provider.modifier.KoVisibilityModifierProvider
@@ -39,6 +39,6 @@ interface KoTypeAliasDeclaration :
     KoSourceSetProvider,
     KoResideInPackageProvider,
     KoTextProvider,
-    KoTypeProvider,
+    KoNonNullableTypeProvider,
     KoVisibilityModifierProvider,
     KoActualModifierProvider
