+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/list/KoNonNullableTypeProviderListExt.kt
@@ -1,13 +1,13 @@
 package com.lemonappdev.konsist.api.ext.list
 
 import com.lemonappdev.konsist.api.declaration.KoTypeDeclaration
-import com.lemonappdev.konsist.api.provider.KoTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
 import kotlin.reflect.KClass
 
 /**
  * List containing type declarations.
  */
-val <T : KoTypeProvider> List<T>.types: List<KoTypeDeclaration>
+val <T : KoNonNullableTypeProvider> List<T>.types: List<KoTypeDeclaration>
     get() = map { it.type }
 
 /**
@@ -16,7 +16,7 @@ val <T : KoTypeProvider> List<T>.types: List<KoTypeDeclaration>
  * @param predicate The predicate function to determine if a declaration type satisfies a condition.
  * @return A list containing declarations with the specified type.
  */
-fun <T : KoTypeProvider> List<T>.withType(predicate: (KoTypeDeclaration) -> Boolean): List<T> =
+fun <T : KoNonNullableTypeProvider> List<T>.withType(predicate: (KoTypeDeclaration) -> Boolean): List<T> =
     filter { predicate(it.type) }
 
 /**
@@ -25,7 +25,7 @@ fun <T : KoTypeProvider> List<T>.withType(predicate: (KoTypeDeclaration) -> Bool
  * @param predicate The predicate function to determine if a declaration type satisfies a condition.
  * @return A list containing declarations without the specified type.
  */
-fun <T : KoTypeProvider> List<T>.withoutType(predicate: (KoTypeDeclaration) -> Boolean): List<T> =
+fun <T : KoNonNullableTypeProvider> List<T>.withoutType(predicate: (KoTypeDeclaration) -> Boolean): List<T> =
     filterNot { predicate(it.type) }
 
 /**
@@ -35,7 +35,7 @@ fun <T : KoTypeProvider> List<T>.withoutType(predicate: (KoTypeDeclaration) -> B
  * @param kClasses The Kotlin class(es) representing the type(s) to include.
  * @return A list containing declarations with the type of the specified Kotlin class(es).
  */
-fun <T : KoTypeProvider> List<T>.withTypeOf(kClass: KClass<*>, vararg kClasses: KClass<*>): List<T> =
+fun <T : KoNonNullableTypeProvider> List<T>.withTypeOf(kClass: KClass<*>, vararg kClasses: KClass<*>): List<T> =
     filter {
         it.hasTypeOf(kClass) ||
             if (kClasses.isNotEmpty()) {
@@ -52,7 +52,7 @@ fun <T : KoTypeProvider> List<T>.withTypeOf(kClass: KClass<*>, vararg kClasses: 
  * @param kClasses The Kotlin class(es) representing the type(s) to exclude.
  * @return A list containing declarations without type of the specified Kotlin class(es).
  */
-fun <T : KoTypeProvider> List<T>.withoutTypeOf(kClass: KClass<*>, vararg kClasses: KClass<*>): List<T> =
+fun <T : KoNonNullableTypeProvider> List<T>.withoutTypeOf(kClass: KClass<*>, vararg kClasses: KClass<*>): List<T> =
     filterNot {
         it.hasTypeOf(kClass) ||
             if (kClasses.isNotEmpty()) {
