+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/list/KoNullableProviderListExt.kt
@@ -1,17 +1,17 @@
 package com.lemonappdev.konsist.api.ext.list
 
-import com.lemonappdev.konsist.api.provider.KoNullableTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNullableProvider
 
 /**
  * List containing declarations with nullable type.
  *
  * @return A list containing declarations that have nullable type.
  */
-fun <T : KoNullableTypeProvider> List<T>.withNullableType(): List<T> = filter { it.isNullable }
+fun <T : KoNullableProvider> List<T>.withNullableType(): List<T> = filter { it.isNullable }
 
 /**
  * List containing declarations without nullable type.
  *
  * @return A list containing declarations that don't have nullable type.
  */
-fun <T : KoNullableTypeProvider> List<T>.withoutNullableType(): List<T> = filterNot { it.isNullable }
+fun <T : KoNullableProvider> List<T>.withoutNullableType(): List<T> = filterNot { it.isNullable }
