+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/provider/KoPropertyTypeProviderExt.kt
@@ -1,8 +1,8 @@
-import com.lemonappdev.konsist.api.provider.KoPropertyTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNullableTypeProvider
 
 /**
  * Determines whatever declaration has type.
  *
  * @return `true` if the declaration has type with the specified KClass name, `false` otherwise.
  */
-inline fun <reified T> KoPropertyTypeProvider.hasTypeOf(): Boolean = hasTypeOf(T::class)
+inline fun <reified T> KoNullableTypeProvider.hasTypeOf(): Boolean = hasTypeOf(T::class)
