+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/ext/provider/KoTypeProviderExt.kt
@@ -1,8 +1,8 @@
-import com.lemonappdev.konsist.api.provider.KoTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
 
 /**
  * Determines whatever declaration has type.
  *
  * @return `true` if the declaration has type with the specified KClass name, `false` otherwise.
  */
-inline fun <reified T> KoTypeProvider.hasTypeOf(): Boolean = hasTypeOf(T::class)
+inline fun <reified T> KoNonNullableTypeProvider.hasTypeOf(): Boolean = hasTypeOf(T::class)
