+++ lib/src/main/kotlin/com/lemonappdev/konsist/api/provider/KoNullableProvider.kt
@@ -1,0 +1,11 @@
+package com.lemonappdev.konsist.api.provider
+
+/**
+ * An interface representing a Kotlin declaration that provides information about whether its type is nullable.
+ */
+interface KoNullableProvider : KoBaseProvider {
+    /**
+     * Whatever type is nullable.
+     */
+    val isNullable: Boolean
+}
