+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/declaration/KoPropertyDeclarationCore.kt
@@ -19,8 +19,8 @@ import com.lemonappdev.konsist.core.provider.KoKDocProviderCore
 import com.lemonappdev.konsist.core.provider.KoLocationProviderCore
 import com.lemonappdev.konsist.core.provider.KoModuleProviderCore
 import com.lemonappdev.konsist.core.provider.KoNameProviderCore
+import com.lemonappdev.konsist.core.provider.KoNullableTypeProviderCore
 import com.lemonappdev.konsist.core.provider.KoPathProviderCore
-import com.lemonappdev.konsist.core.provider.KoPropertyTypeProviderCore
 import com.lemonappdev.konsist.core.provider.KoReceiverTypeProviderCore
 import com.lemonappdev.konsist.core.provider.KoResideInPackageProviderCore
 import com.lemonappdev.konsist.core.provider.KoSetterProviderCore
@@ -75,7 +75,7 @@ internal class KoPropertyDeclarationCore private constructor(
     KoContainingFileProviderCore,
     KoDeclarationFullyQualifiedNameProviderCore,
     KoDelegateProviderCore,
-    KoPropertyTypeProviderCore,
+    KoNullableTypeProviderCore,
     KoInitializerProviderCore,
     KoKDocProviderCore,
     KoLocationProviderCore,
