+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoNonNullableTypeProviderCore.kt
@@ -1,10 +1,10 @@
 package com.lemonappdev.konsist.core.provider
 
 import com.lemonappdev.konsist.api.declaration.KoTypeDeclaration
-import com.lemonappdev.konsist.api.provider.KoTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
 import kotlin.reflect.KClass
 
-internal interface KoTypeProviderCore : KoTypeProvider, KoBaseProviderCore {
+internal interface KoNonNullableTypeProviderCore : KoNonNullableTypeProvider, KoBaseProviderCore {
     override fun hasType(predicate: (KoTypeDeclaration) -> Boolean): Boolean = predicate(type)
 
     override fun hasTypeOf(kClass: KClass<*>): Boolean = kClass.simpleName == type.name
