+++ lib/src/main/kotlin/com/lemonappdev/konsist/core/provider/KoNullableProviderCore.kt
@@ -1,0 +1,11 @@
+package com.lemonappdev.konsist.core.provider
+
+import com.lemonappdev.konsist.api.provider.KoNullableProvider
+import org.jetbrains.kotlin.psi.KtTypeReference
+
+internal interface KoNullableProviderCore : KoNullableProvider, KoBaseProviderCore {
+    val ktTypeReference: KtTypeReference
+
+    override val isNullable: Boolean
+        get() = ktTypeReference.text.last() == '?'
+}
