+++ lib/src/test/kotlin/com/lemonappdev/konsist/api/ext/list/KoNonNullableTypeProviderListExtTest.kt
@@ -1,7 +1,7 @@
 package com.lemonappdev.konsist.api.ext.list
 
 import com.lemonappdev.konsist.api.declaration.KoTypeDeclaration
-import com.lemonappdev.konsist.api.provider.KoTypeProvider
+import com.lemonappdev.konsist.api.provider.KoNonNullableTypeProvider
 import com.lemonappdev.konsist.testdata.SampleType1
 import com.lemonappdev.konsist.testdata.SampleType2
 import io.mockk.every
@@ -9,16 +9,16 @@ import io.mockk.mockk
 import org.amshove.kluent.shouldBeEqualTo
 import org.junit.jupiter.api.Test
 
-class KoTypeProviderListExtTest {
+class KoNonNullableTypeProviderListExtTest {
     @Test
     fun `types returns types from all declarations`() {
         // given
         val type1: KoTypeDeclaration = mockk()
         val type2: KoTypeDeclaration = mockk()
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { type } returns type1
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { type } returns type2
         }
         val declarations = listOf(declaration1, declaration2)
@@ -41,10 +41,10 @@ class KoTypeProviderListExtTest {
         val type2: KoTypeDeclaration = mockk {
             every { name } returns name2
         }
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { type } returns type1
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { type } returns type2
         }
         val declarations = listOf(declaration1, declaration2)
@@ -67,10 +67,10 @@ class KoTypeProviderListExtTest {
         val type2: KoTypeDeclaration = mockk {
             every { name } returns name2
         }
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { type } returns type1
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { type } returns type2
         }
         val declarations = listOf(declaration1, declaration2)
@@ -85,10 +85,10 @@ class KoTypeProviderListExtTest {
     @Test
     fun `withTypeOf(KClass) returns declaration with given return type`() {
         // given
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns true
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
         }
         val declarations = listOf(declaration1, declaration2)
@@ -103,15 +103,15 @@ class KoTypeProviderListExtTest {
     @Test
     fun `withTypeOf(KClass) returns declarations with one of given return types`() {
         // given
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns true
             every { hasTypeOf(SampleType2::class) } returns false
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
             every { hasTypeOf(SampleType2::class) } returns true
         }
-        val declaration3: KoTypeProvider = mockk {
+        val declaration3: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
             every { hasTypeOf(SampleType2::class) } returns false
         }
@@ -127,10 +127,10 @@ class KoTypeProviderListExtTest {
     @Test
     fun `withoutTypeOf(KClass) returns declaration without given return type`() {
         // given
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns true
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
         }
         val declarations = listOf(declaration1, declaration2)
@@ -145,15 +145,15 @@ class KoTypeProviderListExtTest {
     @Test
     fun `withoutTypeOf(KClass) returns declaration without any of given return types`() {
         // given
-        val declaration1: KoTypeProvider = mockk {
+        val declaration1: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns true
             every { hasTypeOf(SampleType2::class) } returns false
         }
-        val declaration2: KoTypeProvider = mockk {
+        val declaration2: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
             every { hasTypeOf(SampleType2::class) } returns true
         }
-        val declaration3: KoTypeProvider = mockk {
+        val declaration3: KoNonNullableTypeProvider = mockk {
             every { hasTypeOf(SampleType1::class) } returns false
             every { hasTypeOf(SampleType2::class) } returns false
         }
