+++ lib/src/test/kotlin/com/lemonappdev/konsist/api/ext/list/KoNullableProviderListExtTest.kt
@@ -1,0 +1,45 @@
+package com.lemonappdev.konsist.api.ext.list
+
+import com.lemonappdev.konsist.api.provider.KoNullableProvider
+import io.mockk.every
+import io.mockk.mockk
+import org.amshove.kluent.shouldBeEqualTo
+import org.junit.jupiter.api.Test
+
+class KoNullableProviderListExtTest {
+    @Test
+    fun `withNullableType() returns type with Nullable basic type`() {
+        // given
+        val type1: KoNullableProvider = mockk {
+            every { isNullable } returns true
+        }
+        val type2: KoNullableProvider = mockk {
+            every { isNullable } returns false
+        }
+        val types = listOf(type1, type2)
+
+        // when
+        val sut = types.withNullableType()
+
+        // then
+        sut shouldBeEqualTo listOf(type1)
+    }
+
+    @Test
+    fun `withoutNullableType() returns type without Nullable basic type`() {
+        // given
+        val type1: KoNullableProvider = mockk {
+            every { isNullable } returns true
+        }
+        val type2: KoNullableProvider = mockk {
+            every { isNullable } returns false
+        }
+        val types = listOf(type1, type2)
+
+        // when
+        val sut = types.withoutNullableType()
+
+        // then
+        sut shouldBeEqualTo listOf(type2)
+    }
+}
